refactor: Refactor MUI-X V6 component tests
Refactors the HTML driver test examples by separating UI markup into “.examples” files and test logic into “.suite” files, enabling the tests to run correctly in Vite.

- Split each example into a UI-only export (`IExampleUIUnit`) and a test suite (`TestSuiteInfo`) in separate `.examples.tsx` and `.suite.ts` files
- Updated all index files and `directory.tsx` to import the new UI-only units, adjusted `ExampleList` to expect `IExampleUIUnit` instead of `IExampleUnit`
