--- packages/component-driver-mui-v6-test/public/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>MUI V6 Test</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
