+++ packages/component-driver-mui-x-v6-test/__tests__/DataGrid.dom.test.ts
@@ -1,5 +1,5 @@
 import { jestTestAdapter } from '@atomic-testing/jest';
-import { createTestEngine } from '@atomic-testing/react';
+import { createTestEngine } from '@atomic-testing/react-19';
 import { testRunner } from '@atomic-testing/test-runner';
 
 import { basicDataGridProExample, basicDataGridProTestSuite } from '../src/examples';
