+++ packages/component-driver-mui-x-v6-test/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>MUI V5 Test</title>
+    <title>MUI X V6 Test</title>
   </head>
   <body>
     <div id="root"></div>
