+++ packages/component-driver-mui-x-v6-test/package.json
@@ -8,8 +8,8 @@
     "src"
   ],
   "scripts": {
-    "start": "PORT=5126 TSC_COMPILE_ON_ERROR=true react-scripts start",
-    "build:ui": "react-scripts build",
+    "start": "vite --force",
+    "build:ui": "vite build",
     "postbuild": "node ../../scripts/postBuild.mjs",
     "test:dom": "jest",
     "test:e2e": "playwright test",
@@ -43,8 +43,8 @@
     "@mui/x-date-pickers-pro": "^6.20.2",
     "dayjs": "^1.11.13",
     "exceljs": "^4.4.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-router-dom": "^7.6.2"
   },
   "devDependencies": {
@@ -55,17 +55,18 @@
     "@atomic-testing/component-driver-mui-v6": "workspace:*",
     "@atomic-testing/component-driver-mui-x-v6": "workspace:*",
     "@atomic-testing/jest": "workspace:*",
-    "@atomic-testing/react": "workspace:*",
+    "@atomic-testing/react-19": "workspace:*",
     "@atomic-testing/playwright": "workspace:*",
     "@atomic-testing/test-runner": "workspace:*",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.15.30",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@vitejs/plugin-react": "^4.5.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^5.8.3"
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5"
   },
   "eslintConfig": {
     "extends": [
