+++ packages/component-driver-mui-x-v6-test/src/examples/datagridpro/index.ts
@@ -1,6 +1,8 @@
+import { JSX } from 'react';
+
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
 
-import { basicDataGridProExample, basicDataGridProTestSuite } from './BasicDataGridPro.examples';
+import { basicDataGridProExample, basicDataGridProTestSuite } from './BasicDataGridPro.suite';
 
 export { basicDataGridProExample, basicDataGridProTestSuite };
 export const dataGridProExamples: IExampleUnit<ScenePart, JSX.Element>[] = [
