+++ packages/component-driver-mui-x-v6-test/vite.config.ts
@@ -1,0 +1,24 @@
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
+
+const port = 5126;
+
+// https://vite.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  resolve: {
+    alias: {
+      '@': '/src', // Assuming your source code is in the 'src' directory
+    },
+  },
+  server: {
+    host: true,
+    port,
+    // Having strictPort set to true along with hmr port locked to the same port
+    // would make sure hot-module-reload works properly
+    strictPort: true,
+    hmr: {
+      port: 5127,
+    },
+  },
+});
